third_party/blink/web_tests/http/tests/websocket/workers/worker-handshake-challenge-randomness-expected.txt

Handshake request in worker should contain random challenge values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/worker-handshake-challenge-randomness.js
PASS [Worker] challenge1 === challenge2 is false
PASS successfullyParsed is true

TEST COMPLETE
